fix(review): clamp the probe timeout, count entries once, register the window as float (#15774, #15777)

+++ autobot_shared/env_registry_logging.py
@@ -61,8 +61,8 @@ register_env_var(
 register_env_var(
     EnvVarSpec(
         name="AUTOBOT_LOG_FLOOD_WINDOW_SECONDS",
-        type=int,
-        default=60,
+        type=float,
+        default=60.0,
         description="Length of the log-flood suppression window, in seconds.",
         component="logging",
     )
